chore: extract shared e2e telemetry checker
