+++ .github/workflows/ci.yml
@@ -239,6 +239,7 @@ jobs:
               - 'e2e/cloud-run.test.ts'
               - 'e2e/helpers/cloud-run-verifier.ts'
               - 'e2e/helpers/cloud-run-telemetry-checker.ts'
+              - 'e2e/helpers/telemetry-checker.ts'
               - '.github/workflows/ci.yml'
 
   check-aas-changes:
